+++ ui/components/icons.mjs
@@ -723,6 +723,15 @@ export const VariableIcon = (props) => (
 )
 
 /*
+ * Windown - Looks like an (application) window
+ */
+export const WindowIcon = (props) => (
+  <IconWrapper {...props}>
+    <path d="M3 8.25V18a2.25 2.25 0 0 0 2.25 2.25h13.5A2.25 2.25 0 0 0 21 18V8.25m-18 0V6a2.25 2.25 0 0 1 2.25-2.25h13.5A2.25 2.25 0 0 1 21 6v2.25m-18 0h18M5.25 6h.008v.008H5.25V6ZM7.5 6h.008v.008H7.5V6Zm2.25 0h.008v.008H9.75V6Z" />
+  </IconWrapper>
+)
+
+/*
  * WifiIcon - Looks like a wifi symbol
  */
 export const WifiIcon = (props) => (