+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- [ui] Add missing inventory/oss icon to main navigation menu
+
 ## [0.6.1] - 2025-01-28
 
 ### Added